Move workbench file dialogues to correct location
