+++ src/serde/workbench.rs
@@ -1,5 +1,3 @@
-use std::path::PathBuf;
-
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -24,16 +22,4 @@ pub struct ProbePin {
 pub struct WorkbenchFile {
     // checksum: (),
     pub probes: Vec<SavedProbe>,
-}
-
-pub fn show_load_workbench_file_dialogue() -> Option<PathBuf> {
-    rfd::FileDialog::new()
-        .add_filter("Cirquil Workbench", vec!["cwb"].as_slice())
-        .pick_file()
-}
-
-pub fn show_save_workbench_file_dialogue() -> Option<PathBuf> {
-    rfd::FileDialog::new()
-        .add_filter("Cirquil Workbench", vec!["cwb"].as_slice())
-        .save_file()
 }
